Use AWS NTP server on EC2 instances
By default, the daemon systemd-timesyncd [1] connects to ntp.ubuntu.com
to synchronize the clock. If the ubuntu server is down, the instance
will stall on this daemon trying to connect to the Ubuntu NTP server,
which will probably result in a tree closure.

We switch to connect to the internal Amazon NTP server [2] to avoid this
potential outage.

[1] https://wiki.archlinux.org/index.php/systemd-timesyncd
[2] https://aws.amazon.com/blogs/aws/keeping-time-with-amazon-time-sync-service/